# docs/swagger.yaml
basePath: /
definitions:
  handlers.ApplyPatchRequest:
    properties:
      patch:
        type: string
      path:
        type: string
    type: object
  handlers.CommitResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      snapshot_id:
        type: string
      version:
        type: integer
    type: object
  handlers.DiffResponse:
    properties:
      diff:
        type: string
      path:
        type: string
      version:
        type: integer
    type: object
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  handlers.FileEntry:
    properties:
      is_dir:
        type: boolean
      mod_time:
        type: string
      mode:
        type: integer
      path:
        type: string
      size:
        type: integer
    type: object
  handlers.ListResponse:
    properties:
      entries:
        items:
          $ref: '#/definitions/handlers.FileEntry'
        type: array
      path:
        type: string
    type: object
...
  handlers.ReadResponse:
    properties:
      content:
        type: string
      encoding:
        type: string
      mod_time:
        type: string
      mode:
        type: integer
      path:
        type: string
      size:
        type: integer
    type: object
  handlers.WriteAtomicRequest:
    properties:
      content:
        type: string
      encoding:
        type: string
      mode:
        type: integer
      mtime:
        type: string
      path:
        type: string
    type: object
...
info:
  contact: {}
  description: User-scoped filesystem API for containerd-backed data.
  title: Memoh Go API
  version: "1.0"
paths:
...
  /fs/apply_patch:
    post:
      description: Apply a unified diff patch to a file under the user data mount
      parameters:
      - description: Patch payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.ApplyPatchRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Apply unified diff patch
      tags:
      - fs
  /fs/commit:
    post:
      description: Create a new version snapshot for the user container
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CommitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Commit a filesystem snapshot
      tags:
      - fs
  /fs/diff:
    get:
      description: Produce a unified diff between a version snapshot and current data
      parameters:
      - description: Path under data mount
        in: query
        name: path
        type: string
      - description: Version number
        in: query
        name: version
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DiffResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Diff against a version snapshot
      tags:
      - fs
  /fs/list:
    get:
      description: List files under the user data mount
      parameters:
      - description: Path under data mount
        in: query
        name: path
        type: string
      - description: Recursive listing
        in: query
        name: recursive
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List directory contents
      tags:
      - fs
  /fs/read:
    get:
      description: Read a file under the user data mount
      parameters:
      - description: Path under data mount
        in: query
        name: path
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ReadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Read file content
      tags:
      - fs
  /fs/write_atomic:
    put:
      description: Atomically replace a file under the user data mount
      parameters:
      - description: Write payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.WriteAtomicRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Write file atomically
      tags:
      - fs
...
schemes:
- http
swagger: "2.0"